lambda_functions/nba_scraper/src/requirements.txt
requests
botocore~=1.34.20
selenium~=4.18.1
nba_api~=1.4.1
beautifulsoup4~=4.12.2
bs4~=0.0.1
python-dotenv~=0.21.1
tqdm~=4.66.1
boto3~=1.34.20
webdriver-manager~=4.0.1
openai~=1.6.1
GitPython~=3.1.42